move files to src/pages
